pyproject.toml
[tool.poetry]
name = "oop"
version = "0.1.0"
description = ""
authors = ["aleksejbelkin1017 <aleksejbelkin1017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# Максимальная длина строки
line_length = 119